Table/Tree: do nothing if no elements are given on insert/update/delete
Use case: editing a smart column sends en empty update event because the
table update buffer calls updateRows even if the rows array is empty.
Instead of fixing this in the update buffer the table and tree are
change to prevent it for other cases too.

This makes it consistent with the tile grid.