ContextMenu: add animation
Adding animations leads to following problems:
- When the user closes the context menu while a sub menu is opening,
  an exception occurs. This happens because the scrollbars, that should
  be uninstalled during the complete callback, won't be uninstalled
  because the removal is pending. But the data('scrollbars') is removed,
  so once the removal is done and the children (=scrollbars) should
  really be removed, it is not possible anymore, because data is
  missing.
  The solution is to not call complete() when it is not rendered (or
  removal is pending), see ContextMenuPopup.js
  Also, maybe scrollbars.uninstall should be more failsafe, but this
  has not changed.
- If the submenu is clicked during opening animation, the size of the
  popup won't be correct -> delay the submenu click until opening is
  done, see ContextMenuPopup.js
- If the context menu contains a form menu it is important to remove
  the form immediately as soon as that form should be shown again
  in another menu. Even though menu items are cloned, forms are not ->
  already rendered exceptions. See FormMenu.js

Also made sure a menu can open the popup again when it is clicked
while the popup is being removed. This is necessary to always have a
consistent state between menu and popup.

Also improved FormMenu support for Scout JS apps by setting necessary
properties (see FormMenu.adaptForm) and making sure, state is correct
after form destroyed (now behaves similar as FormTableControl).
The previous logic (remove listener) did interfere with the new
possibility to reopen the popup while being removed. But actually
doing it only on destroy is more correct anyway because remove may
only be temporary and should not modify model state. This way the popup
can be opened again if it was only removed temporarily (e.g. a view
containing a menu with an open form is detached and attached again).