BrushUp: improve focus style of checkable table
Until now, if a checkable table uses checkableStyle = TABLE_ROW,
a focus indicator is shown on the beginning of the selected row.
This works but does not look very good.
To make it prettier, the table now uses the new possibility to only
visualize the focus marker when the user uses the keyboard. Now, the
focus marker can be replaced by a better visualization which would
be confusing when the user uses a mouse but works well if he uses the
keyboard.