Mobile: fix error when closing form
Since commit 39192bb0e5bc5c5e03f174c31765e543d33f52ef
the display parent is set for detail forms. When the
display parent is set, the parent will be adjusted
automatically using Desktop.computeParentForDisplayParent.
In our case, parent is set to desktop because display
parent is outline.
This breaks the mobile client because the detail form is not removed
when a regular form is opened which leads to exceptions when the form
is closed again.
Actually, detail forms don't need a display parent at all.